@@ -13,10 +13,10 @@
     ],
     "require": {
         "php": "^7.1",
-        "symfony/framework-bundle": "^3.4 || ^4.0",
-        "symfony/debug": "^3.4 || ^4.0",
+        "symfony/framework-bundle": "^4.0 || ^5.0",
+        "symfony/debug": "^4.0 || ^5.0",
         "doctrine/common": "^2.3",
-        "symfony/process": "^3.4 || ^4.0"
+        "symfony/process": "^4.0 || ^5.0"
     },
     "require-dev": {
         "doctrine/doctrine-fixtures-bundle": "*",
